Warn if discarding or no changes
